Fix fileprovider build with androidx
When androidx is enabled, the android support library is no longer
available. The library artifact needs to be replaced as well as changing
code to use the `FileProvider` class from `androidx.core.content`
instead of `android.support.v4.content`.

Unfortunately, multidex is mixed up in this even though it doesn't have
anything to do FileProvider or general usage of androidx. Keep them
together for now.